Convert to/hold our path in a useful format as early as possible

+++ app/src/main/java/com/greenaddress/greenapi/ISigningWallet.java
@@ -20,7 +20,7 @@ public interface ISigningWallet {
 
     DeterministicKey getPubKey();
 
-    List<ECKey.ECDSASignature> signTransaction(PreparedTransaction ptx, byte[] gaitPath);
+    List<ECKey.ECDSASignature> signTransaction(PreparedTransaction ptx, int[] gaitPath);
 
     boolean requiresPrevoutRawTxs();
 }
